refactor(ground-truth): move ground_truth module from extraction.evaluation.benchmark to core
Relocates ground_truth.py to src/core/ and updates all imports across evaluation, predictions, runner, scripts, and tests to reflect the new path.
