+++ tests/test_groundwater.py
@@ -5,7 +5,7 @@ from datetime import date
 import pytest
 
 from core.benchmark_utils import Metrics
-from extraction.evaluation.benchmark.ground_truth import GroundTruth
+from core.ground_truth import GroundTruth
 from extraction.evaluation.groundwater_evaluator import (
     GroundwaterEvaluator,
 )
