+++ tests/test_predictions.py
@@ -7,7 +7,7 @@ import pymupdf
 import pytest
 
 from core.benchmark_utils import Metrics
-from extraction.evaluation.benchmark.ground_truth import GroundTruth
+from core.ground_truth import GroundTruth
 from extraction.evaluation.benchmark.metrics import OverallMetricsCatalog
 from extraction.evaluation.layer_evaluator import LayerEvaluator
 from extraction.evaluation.utility import evaluate, evaluate_single
